chop d_lightstylevalue from 8.8 to 2.6 to fit in byte

+++ source/r_main.c
@@ -107,7 +107,7 @@ texture_t	*r_notexture_mip;
 
 float		r_aliastransition, r_resfudge;
 
-int		d_lightstylevalue[256];	// 8.8 fraction of base light value
+byte		d_lightstylevalue[256];	// 2.6 fraction of base light value
 
 float	dp_time1, dp_time2, db_time1, db_time2, rw_time1, rw_time2;
 float	se_time1, se_time2, de_time1, de_time2, dv_time1, dv_time2;
